+++ back-end/routers/products.js
@@ -2,6 +2,7 @@ const { Product } = require("../models/product");
 const { Category } = require("../models/category");
 const express = require("express");
 const router = express.Router();
+const mongoose = require("mongoose");
 
 router.get(`/`, async (req, res) => {
   const productList = await Product.find().populate("category");
@@ -52,6 +53,10 @@ router.post(`/`, async (req, res) => {
 });
 
 router.delete("/:id", (req, res) => {
+  if (!mongoose.isValidObjectId(req.params.id)) {
+    return res.status(400).send("Invalid Product Id");
+  }
+
   Product.findByIdAndRemove(req.params.id)
     .then((product) => {
       if (product) {
@@ -70,19 +75,31 @@ router.delete("/:id", (req, res) => {
 });
 
 router.put("/:id", async (req, res) => {
-  const product = await Product.findByIdAndUpdate(req.params.id, {
-    name: req.body.name,
-    description: req.body.description,
-    richDescription: req.body.richDescription,
-    image: req.body.image,
-    brand: req.body.brand,
-    price: req.body.price,
-    category: req.body.category,
-    countInStock: req.body.countInStock,
-    rating: req.body.rating,
-    numberReviews: req.body.numberReviews,
-    isFeatured: req.body.isFeatured,
-  });
+  if (!mongoose.isValidObjectId(req.params.id)) {
+    return res.status(400).send("Invalid Product Id");
+  }
+
+  const category = await Category.findById(req.body.category);
+
+  if (!category) return res.status(400).send("Invalid Category");
+
+  const product = await Product.findByIdAndUpdate(
+    req.params.id,
+    {
+      name: req.body.name,
+      description: req.body.description,
+      richDescription: req.body.richDescription,
+      image: req.body.image,
+      brand: req.body.brand,
+      price: req.body.price,
+      category: req.body.category,
+      countInStock: req.body.countInStock,
+      rating: req.body.rating,
+      numberReviews: req.body.numberReviews,
+      isFeatured: req.body.isFeatured,
+    },
+    { new: true }
+  );
 
   if (!product) return res.status(401).send("the product cannot be update!");
 
@@ -93,12 +110,23 @@ router.get("/:id", async (req, res) => {
   const product = await Product.findById(req.params.id).populate("category");
 
   if (!product) {
-    res
-      .status(500)
-      .json({ message: "The product with the given ID was not found!" });
+    res.status(500).json({
+      success: false,
+      message: "The product with the given ID was not found!",
+    });
   }
 
   res.status(200).send(product);
+});
+
+router.get("/get/count", async (req, res) => {
+  const productCount = await Product.countDocuments();
+
+  if (!productCount) {
+    res.status(500).json({ success: false });
+  }
+
+  res.send({ productCount: productCount });
 });
 
 module.exports = router;
